pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-clipper"
version = "0.1.0"
authors = [
  { name="frostmute" },
]
description = "Agent-driven Obsidian web clipper powered by Hermes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "beautifulsoup4",
    "readability-lxml",
    "watchdog",
    "fastapi",
    "uvicorn",
    "pydantic",
    "markdownify",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
hermes-clip = "hermes_clipper.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["hermes_clipper*"]
